# lib/api-zod/src/generated/types/locationCheckInInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CouplesBudget API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LocationCheckInInput {
  lat: number;
  lng: number;
  /** @nullable */
  locationName?: string | null;
}